move default implementation of json-like
